# app/openapi.yaml
openapi: 3.0.3
info:
  title: WhatsApp Multi-Instance API
  description: |
    RESTful API for managing multiple WhatsApp instances.

    **Base URL:** use the host you opened this spec on (each Azure region is its own base URL, e.g. `https://wasup-uk-south.azurewebsites.net`). The Scalar page rewrites `servers` automatically.

    ## Authentication
    When authentication is enabled, **every** `/api/...` request must include a valid org API key from the Wasup dashboard Connection page.

    Use one of:
    - Header: `X-API-Key: <your-org-api-key>`
    - Header: `Authorization: Bearer <your-org-api-key>`

    Example (replace `BASE` and `KEY` with this region's URL and your org key):

    ```bash
    curl -sS -H "X-API-Key: KEY" "BASE/api/instances"
    ```

    ```bash
    curl -sS -X POST "BASE/api/instances/INSTANCE_ID/send" \
      -H "X-API-Key: KEY" -H "Content-Type: application/json" \
      -d '{"to":"447700900000@s.whatsapp.net","message":"Hello"}'
    ```

    ## Interactive buttons
    Custom quick-reply buttons are supported on the same send endpoints. Set
    `messageType: buttons`, or omit `messageType` and include a non-empty
    `buttons` array. Use `message` or `text` for the body.

    ```bash
    curl -sS -X POST "BASE/api/instances/INSTANCE_ID/send" \
      -H "X-API-Key: KEY" -H "Content-Type: application/json" \
      -d '{"to":"60123456789","text":"Choose an option","buttons":[{"id":"yes","text":"Yes"},{"id":"no","text":"No"}],"footer":"Wasup"}'
...

...

...

...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/instances/{instanceId}/clear-auth:
    parameters:
      - $ref: '#/components/parameters/instanceId'
    post:
      tags: [Connection]
      summary: Clear auth
      description: Logout and delete saved credentials. Requires new QR scan.
      responses:
        '200':
          description: Auth cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Auth cleared
                  instance:
                    $ref: '#/components/schemas/Instance'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/instances/{instanceId}/handoff:
    parameters:
      - $ref: '#/components/parameters/instanceId'
    get:
      tags: [Handoff]
      summary: Get active human-handoff chats
      description: |
        Returns the list of chat JIDs that are currently in human-mode
        (bot responses are paused). A chat enters human-mode when you send
        a manual message from the phone or via this API. Resume bot mode by
        sending a message containing `#ai`, `#assistant`, `#bot`, or `#resume`.
      responses:
        '200':
          description: Handoff list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  chats:
                    type: array
                    items:
                      type: object
                      properties:
                        jid: { type: string, example: '447835156367@s.whatsapp.net' }
                        since: { type: string, format: date-time }
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Handoff]
      summary: Tag or untag a chat for human handoff
      description: |
        Manually pause or resume bot responses for a specific phone number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, active]
              properties:
                phone:
                  type: string
                  description: Phone number or JID to tag
                  example: '447835156367'
                active:
                  type: boolean
                  description: '`true` to pause bot, `false` to resume'
                  example: true
      responses:
        '200':
          description: Handoff updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  jid: { type: string }
                  active: { type: boolean }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Handoff]
      summary: Clear all human handoff tags
      description: Resumes bot mode for all chats on this instance.
      responses:
        '200':
          description: All handoffs cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  cleared: { type: integer, example: 3 }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    instanceId:
      name: instanceId
      in: path
      required: true
      schema:
        type: string
      description: Instance ID
      example: wa_lxyz123_abc45
  schemas:
...
    Instance:
      type: object
      properties:
        id:
          type: string
          example: wa_lxyz123_abc45
        name:
          type: string
          example: Customer Support
        status:
          type: string
          enum: [disconnected, connecting, connected]
          example: connected
        qrCode:
          type: string
          nullable: true
          description: Base64 QR code (only when connecting)
        pairingCode:
          type: string
          nullable: true
          description: Pairing code (only when connecting via pairing)
        connectedPhone:
          type: string
          nullable: true
          example: '60123456789'
        connectedAt:
          type: string
          format: date-time
          nullable: true
        webhookUrl:
          type: string
          example: https://your-api.com/webhook
        behaviorSettings:
          $ref: '#/components/schemas/BehaviorSettings'
        antiBanSettings:
          $ref: '#/components/schemas/AntiBanSettings'
        antiBanHealth:
          $ref: '#/components/schemas/AntiBanHealth'
        createdAt:
          type: string
          format: date-time
    BehaviorSettings:
      type: object
      properties:
        behaviorProfile:
          type: string
          enum: [bot-native, notification-balanced, notification-max]
          description: |
            Top-level behavior mode.

            `bot-native` uses active-device automation with typing, reads, delays,
            and presence cycling.

            `notification-balanced` prioritizes handset alerts without promising
            delivery: inbound messages are logged/webhooked immediately, then
            reads/typing/replies wait for `notificationGraceMs`.

            `notification-max` maximizes handset alerts: typing is disabled and
            automated read receipts are skipped.
          default: bot-native
          example: notification-balanced
        typingSimulation:
          type: boolean
          description: |
            Show "typing..." before sending. In `notification-balanced`, typing
            happens only after the grace window. In `notification-max`, this is
            forced off.
          default: true
          example: true
        delayEnabled:
          type: boolean
          description: Add human-like delays before responding
          default: true
          example: true
        notificationGraceMs:
          type: integer
          minimum: 0
          maximum: 120000
          description: |
            Milliseconds to wait after an inbound message before read receipt,
            typing, or reply in notification profiles. The webhook/log event is
            still sent immediately.
          default: 12000
          example: 12000
        phoneNotificationsEnabled:
          type: boolean
          description: |
            Legacy compatibility flag. New clients should set `behaviorProfile`.
            When true, maps to `notification-balanced` unless typing is false,
            in which case it maps to `notification-max`.
          default: false
          example: false
    AntiBanSettings:
      type: object
      properties:
        preset:
          type: string
          enum: [conservative, balanced, aggressive, custom]
          example: balanced
        messagesPerHour:
          type: integer
          example: 200
        messagesPerDay:
          type: integer
          example: 5000
        uniqueChatsPerHour:
          type: integer
          example: 50
        uniqueChatsPerDay:
          type: integer
          example: 500
    AntiBanHealth:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, warning, limited]
          example: healthy
        hourlyUsage:
          type: integer
          description: Percentage of hourly message limit used
          example: 25
        dailyUsage:
          type: integer
          description: Percentage of daily message limit used
          example: 10
        hourlyChatsUsage:
          type: integer
          example: 20
        dailyChatsUsage:
          type: integer
          example: 8
        warnings:
          type: array
          items:
            type: string
          example: []
        stats:
          type: object
          properties:
            messagesThisHour:
              type: integer
              example: 12
            messagesThisDay:
              type: integer
              example: 30
            uniqueChatsThisHour:
              type: integer
              example: 5
            uniqueChatsThisDay:
              type: integer
              example: 8
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Instance not found
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...